Changes to direct expansion, collusion check.
Changed direct expansion to take a 'step size' parameter so that
direct expansion expands to goal incrementally up until step size. If
robot doesn't collide with obstacle in that path, then that path is
returned, otherwise, no path is returned. Previous version was cutting
the path 'step size' amount after the path is found. So this change
easens the computational time a little.

Removed overestimation (making obstacles bigger than they are) from
collusion check function to the script to again, easen the computation.