Addition of intermediary path in-between nodes to both RRT and RRT*.
- Now every node also stores the direct path to its parent node. This
might seem unneccessary as node already stores its parent node so this
path can be computed afterwards again, which would save space. However,
there appeared to be a problem with that where the addition of a new
node to the tree is done with step size, when this is done, a new node
is an intermediary node to go from already existing node to the randomly
generated one. Suppose that the direct path to go from N (tree node) to
 R (random one) is 'path' and step size is step_size. Then there is no guarantee that
 directPath(N, step_size, path{i}) and directPath(N, step_size, R) is the same.
 This is again, solved by attaching the path to the node.